+++ packages/client/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Storage `watchEntries` API
 - **CLI**:
   - Option flag `noDescriptorsPackage` to generate descriptors without installing them as a package.
   - Add version option